update frame producer
