+++ app/src/main/java/com/zcc/mediarecorder/demo/Camera1GLSurfaceRender.java
@@ -9,8 +9,8 @@ import android.os.Handler;
 import android.os.Looper;
 
 import com.zcc.mediarecorder.demo.utils.Camera1Manager;
-import com.zcc.mediarecorder.gles.FullFrameRect;
-import com.zcc.mediarecorder.gles.Texture2dProgram;
+import com.zcc.mediarecorder.frameproducer.gles.FullFrameRect;
+import com.zcc.mediarecorder.frameproducer.gles.Texture2dProgram;
 
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
