+++ mediarecorderlib/src/main/java/com/zcc/mediarecorder/encoder/core/IVideoEncoderCore.java
@@ -1,0 +1,15 @@
+package com.zcc.mediarecorder.encoder.core;
+
+import android.view.Surface;
+
+import com.zcc.mediarecorder.common.ILifeCircle;
+
+public interface IVideoEncoderCore extends ILifeCircle {
+
+    public Surface getInputSurface();
+
+    public void drainEncoder(boolean endOfStream);
+
+    public long getPTSUs();
+
+}
