+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':mediarecorderlib', ':mediadecorderlib'
+include ':app', ':mediarecorderlib', ':mediadecorderlib', ':glrender'
